First version
This is the first version of the app. It's not too bad, because I learned a lot while trying to put it together. But it's still pretty basic. It has no tests whatsoever. It groups files together in directories according to the type of file (all the classes are bundled together, and all the interfaces as well, etc).
The whole thing chokes if armies are assembled with more than 500 unites per army. That's because I didn't know how to optimize the way that units from opposing armies attack each other.